added winner check
added round ++ so each round will be added by one if there is no
winner. if there is a winner kept the alert up to continue the loop
until there is one or they both die. Then added a break if anyone one
of those conditions  occur to stop running the program